Improve token selection and add support for copy/paste token using buttons (#48)
* fix native auth text input selection

* fix scrollbar

* fix nativ auth input selection and make it more restricted

* fix keys combinations conditions

* big refactoring in order to have paste token functionality in place

* adjust copy and paste buttons

* refactoring

* cleanup
