+++ src/pages/Authentication/components/Input/components/Textarea/types.ts
@@ -1,9 +1,7 @@
 import type { Dispatch, SetStateAction } from 'react';
-import type { FormikProps } from 'formik';
 import type { MetricType } from 'pages/Authentication/types';
-import type { FormValuesType } from '../../types';
 
-export interface TextareaPropsType extends FormikProps<FormValuesType> {
+export interface TextareaPropsType {
   setMetrics: Dispatch<SetStateAction<MetricType>>;
 }
