+++ src/pages/Authentication/context/AuthenticationProvider.tsx
@@ -1,0 +1,19 @@
+import { ReactNode } from 'react';
+import {
+  AuthenticationContext,
+  AuthenticationContextType
+} from './AuthenticationContext';
+
+export const AuthenticationProvider = ({
+  value,
+  children
+}: {
+  value: AuthenticationContextType;
+  children: ReactNode;
+}) => {
+  return (
+    <AuthenticationContext.Provider value={value}>
+      {children}
+    </AuthenticationContext.Provider>
+  );
+};
