+++ src/pages/Authentication/context/index.ts
@@ -1,0 +1,6 @@
+export {
+  AuthenticationContext,
+  useAuthenticationContext,
+  useAuthenticationValue
+} from './AuthenticationContext';
+export { AuthenticationProvider } from './AuthenticationProvider';
